+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - Stirring now has the ability to restart itself if it detects it has stalled.
  - od_normalization has been removed. Instead, there is a small routine `od_statistics` that is called by both `od_blank` and `growth_rate_calculating`. The latter also now stores the required od statistics to storage (previous it was the `od_normalization` job).
  - The LEDs in the pioreactor have been moved up 0.5mm.
+ - Error-handling and user improvements to `pump_calibration`
 
 
 ### 22.7.0